requirements.txt
# TODO: update requirements
fastapi
pydantic
requests
uvicorn
gunicorn


google-cloud-vision
numpy
opencv-python  # cv2
joblib
pdf2image

ipykernel

black
flake8
pytest
